requirements.txt
backoff==1.8.0
certifi==2022.12.7
charset-normalizer==3.0.1
ciso8601==2.3.0
idna==3.4
jsonschema==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
simplejson==3.11.1
singer-python==5.13.0
six==1.16.0
urllib3==1.26.14
